finalni

+++ webapi/Models/Ucesnik.cs
@@ -9,7 +9,7 @@ namespace Models {
         [DefaultValue(0)]
         public int Mesto { get; set; }
         [DefaultValue(0)]
-        public int Bodovi { get; set; }
+        public float Bodovi { get; set; }
         public int TurnirID { get; set; }
         public Turnir Turnir { get; set; }
         public int IgracID { get; set; }
